mac: Fix Wifi_MACWrite()
This function didn't make sense. The function is meant to be used to
write data to MAC RAM to the TX buffer. However, this buffer isn't
circular. The hardware doesn't wrap around when it reaches the end.

DSWiFi simply writes everything to be sent to the start of the TX buffer
in MAC RAM, so the best thing this function can do is check if we're
overflowing that buffer.

Bug reported by nocash in the source code of wifiboot (bug reports
compiled by asie):

https://problemkaputt.de/wifiboot.htm