frame: Generate Null data frames correctly
The IEEE 802.11 frame header was incorrect. It was missing the address 3
and the sequence control fields. Also, address 1 was incorrectly set to the
AP MAC address, but according to "IEEE 802.11 handbook" Table 3-2, when
ToDS=1 and FromDS=0, address 1 is BSSID, address 2 is SA, and address 3 is
DA.

The sizes were also calculated incorrectly.