frame: Refactor data rate handling
The old code got the list of rates supported by the AP and it added the two
rates of the NDS (1 and 2 Mbit/s) at the start of that list if they weren't
found in the original list. This resulted in 1 and 2 being at the start,
and a list of unsupported rates until the limit of the frame (up to 6 rates
that were in the original list because there is a limit of 8 rates).

This is very unusual behaviour, so I've tested a bit, and it looks like
modern routers really don't like it if you only advertise 1 and 2 when you
try to connect. They return error 18 "The mobile station [the NDS] doesn't
support all the data rates required by the BSS [Basic Service Set]".

So the original code was meant to trick APs into accepting us. However,
this combination of rates is strange, and it probably isn't required.
Realistically, for a router to accept us, it needs to support 802.11b.

The new code tries to access twice. The first time it only advertises
the 2 rates supported by the NDS. The second time it advertises all 4
rates of the 802.11b standard. This seems to be at least as reliable as
the previous code.