frame: Refactor code that tracks the size of frames
Previously, Wifi_GenMgtHeader() returned an index to the first byte in the
buffer that you could edit (after the two headers and potentially the WEP
IV+WEP field). Now, it simply returns the number of bytes used of the body
(so 4 bytes if WEP is enabled, 0 bytes if WEP is disabled).

Also, the sizes calculated at the end of the Wifi_Send*Packet() functions
now are more intuitive. Previously they were calculated by adding and
subtracting values. With this patch, the code only adds values in a clean
way.