chore: Remove some commented-out code
The code in Wifi_RxSetup() is definitely not needed, it's just a switch to
configure different ranges for the RX buffer. This library only has one
unique setting regardless of the mode.

The code in Wifi_TxSetup() seems to be modifying the frame body of some
packets already in MAC RAM. This isn't needed in this library.