rx: Refactor Wifi_sgIpHandlePackage()
The old code was a bit messy because it was originally designed to handle
different data for unencrypted and WEP-encrypted packets. However, this is
not required because the hardware removes all additional fields added in
WEP-encrypted packets.

Also, Wifi_sgIpHandlePackage() now doesn't receive the hardware RX header,
only the IEEE 802.11 header and data.