wep: Cleanup code that copied WEP keys around
Sometimes 20 bytes were copied, other times it was 16. The "wepkey7" and
"wepkey9" arrays were 20 bytes long, and "W_WEPKEY_*" is 32 bytes long, so
the destination was always big enough to hold the result, but it doesn't
make sense to copy more than required. For example, Wifi_ConnectAP() would
copy bytes from the "wepkey" argument even after the end of the key,
filling the destination array with garbage. The same thing happened to keys
stored in the firmware.

This commit fixes this situation by always copying the right amount of
bytes.