update: Retry from the start when connecting to an AP
When the connection process gets stuck (particularly, after authenticating
with the AP), it is very likely that retrying just the association step
won't resolve the issue. It's better to start always from the start. When
it works, the process is so fast that it doesn't make any difference from
the user's perspective.