chore: Remove duplicated structs, enums and defines
Some things were defined in public and private headers. We need to remove
the extra copy to prevent issues in the future, like changing only one
copy and forgetting about the other one.

In fact, WIFIINIT_OPTION_USEHEAP_128 and WIFIINIT_OPTION_USEHEAP_64 had two
sets of definitions.

It's better to only keep them in the public headers as they are documented
better there.