mp: Keep track of connected guests
Instead of blindly accepting association requests, keep track of which
guests have connected to the host. That way guests can go from disconnected
to authenticated and finally to associated.

Also, a new guest limit has been added. It is needed to track the MAC
address and association ID of multiple guests, so it is required to limit
the number of devices we keep track of. The default limit is 15, but it
will be possible to change it in the future.