scan: Fix channel change during scan mode
Commit 6edc0df1686b ("ipc: Let users change channel again") broke scan
mode. The ARM9 sets reqChannel and the ARM7 reads that to update the
channel. However, the ARM7 also uses that variable to iterate through
all channels. Instead of that, the ARM7 now sets the channel directly
from the scan mode code, but it also sets reqChannel so that after scan
mode the last channel remains.