ipc: Add global variable to save the current library mode
Sometimes it is needed to know if the library is acting as a multiplayer
host, multiplayer client or accessing the Internet. Until now it was needed
to check "curMode == WIFIMODE_ACCESSPOINT" to see if it was in multiplayer
host mode, and to check the value of "ap_is_multiplay_host" to determine
if it was in multiplayer client or in internet access modes.

Instead of that, now there is a new variable ("curLibraryMode") which
contains the current mode so that it's easier to check that.