scan: Add option to filter out APs in scan mode
It is now possible to filter out some APs when in scan mode. It is possible
to request to only include compatible (and/or incompatible) APs in the
list. It is also possible to ask it to list APs with the Nintendo vendor
information.

For example, a multiplayer client may choose to only list NDS devices
acting as hosts. A NDS trying to access the Internet may only list
compatible APs (to not overwhelm the user) or it may choose to list
incompatible APs as well (to show the user that their AP isn't compatible).