mac: Move memory map defines to a common ARM9/ARM7 header
The sizes need to be used on the ARM9 to check if beacon and client
response frames fit before even trying to send them to the ARM7.

Also, this commit reduces a bit the TX buffer size to add room for a new
client reply. Note that the new TX buffer size is still bigger than the
original pre-multiplayer size (the space reserved for the beacon was
much bigger, which isn't really needed).