tx: Use status field instead of rate to flag CMD packets
The status field of the TX hardware header needs to be set to 0 when a
packet is sent, so it is easier to use this field for DSWifi flags than
the transfer rate.

Also, the transfer rate only needs to be set in packets sent with LOC3. All
CMD packets need to be sent with a 2 MBit/s rate.