mp: Optimize handling of REPLY frames
Instead of copying them to the regular TX buffer and then copying them to
their real destination, do it right away.

Also, remove wifi_lastreply, which kept track of which one of the two reply
buffers was supposed to be used next. It's better to always use the
opposite buffer to the one being used by the W_TXBUF_REPLY2 register (the
one being actively transferred).