mp: Implement user handlers of CMD and REPLY packets
Now it is possible for user code to easily handle CMD and REPLY packets
without having to use the raw received packet handler. The new handlers
only let you read the user data, skipping the IEEE 802.11 header. It
also does some checking to ensure that the packet is actually from a
known host/client, making it so that the developer doesn't have to do
it in the application code.