chore: Remove timeout counter field from client struct
This is unused, and there is no need for DSWifi to truly determine whether
a client is alive or not. The game could have crashed and the wifi chip
would still send empty MP replies. Even if the game actually sends data
replies, only the game code can determine if the client is behaving in a
logical way or not. For this reason, it's better to leave the job of
keeping a timeout counter to the game code.