mp: Allow sending arbitrary data frames between host and clients
Using the CMD/REPLY is very useful for games with many clients, but it can
be unhelpful for 2-player games that transfer varying amounts of data every
time. In that case, it's better to have a way to transfer arbitrary data
packets, even if there is no automated way to reply to them from the
client.

Two new functions have been added:

- Wifi_MultiplayerHostToClientDataTxFrame()
- Wifi_MultiplayerClientToHostDataTxFrame()

The WifiFromHostPacketHandler and WifiFromClientPacketHandler types have
been modified to have an additional argument that says whether the
original packet was a CMD/REPLY packet or a regular data packet.