ap: Make scan mode more consistent
Including channels 1, 6 and 11 several times in the list of channels to
scan skews the values of the timeout counters in the AP structs. It's
better to scan all channels the same number of times so that the APs of
some channel don't get penalized. This penalization becomes clear when a
new AP is added to the list of found APs and we need to remove an AP that
is currently in the list.