ap: Stop keeping track of old RSSI values
In practice, the filter didn't work. Normally, in scan mode, the RSSI value
of an AP would change a lot in a second when the channel of that AP is
selected. Then, when the channel changes, it would stop changing. This
means that all the values would be received in such a short period of time
that the filter was just applied to a lot of values that are very similar.

It's better to save the RAM and remove all the space used by old RSSI
values.