ipc: Implement function to deinitialize DSWiFi
The system is still buggy, but it's a start. sgIP can't be safely
deinitialized and reinitialized, so only local multiplayer mode works
after a restart.

For now, we keep sgIP initialized even after a Wifi_Deinit() so that it
works after DSWiFi is re-initialized. This will keep the sgIP heap in
RAM but, at least, it will free the DSWiFi IPC memory and stop using the
DSWiFi timer.