sgIP: Cleanup of DNS code
The variables that were declared as "volatile" didn't actually need it.

All private variables used in the file have been made static.

All variables are now initialized when sgIP is initialized, except for the
ones that are initialized by sgIP_DNS_GenerateHostent(). They can't be
accessed before the function is called, and it initializes all of them.