sgip: lwip: Replace sgIP by lwIP
General changes:

- Some details of the documentation have been updated.
- The public headers have been updated to match the definitions from
  lwIP headers.
- All pre-existing features are still supported. The only exception is
  `Wifi_Deinit()`, which is currently unsupported if lwIP has been
  initialized.

The result is that DSWifi now has a more modern and widespread IP stack
that we don't need to maintain ourselves. It can also easily support
protocols like IPv6 in the future.

The main issue is that lwIP uses a fair bit more of code than sgIP, and
that programs need to be aware of threading and yield every now and then
to give lwIP a chance to update its state.