twl: Make card init asynchronous
wifi_card_init() may be called from interrupt handlers, but it has a busy
wait loop at the end of the function. This loop only finishes if a message
is received from the card and the interrupt is handled, but that can't
happen while inside an interrupt handler safely.

Instead, the loop has been removed and a new WIFIMODE_INITIALIZING enum
value has been added, which waits until the card is ready.

The call to wmi_scan() at the end of wifi_card_init() has also been removed
and moved to Wifi_TWL_Update() so that it is called only when entering
mode WIFIMODE_SCAN.