ap: Add a new security type public field to AP struct
Instead of using the WEP and WPA flags it is now possible to use the new
field security_type to determine the security type of the network.

The internal code of the NTR driver has been modified to use a
Wifi_ApSecurityType variable instead of individual wepmode and wpamode
variables for simplicity.