ap: Remove ad hoc AP support
This wasn't tested, probably incomplete, it wasn't supported in the DSi
driver, and it was making the code more complicated for no gain (as
nobody requires support for this kind of APs at the moment).

The public flag in "dswifi_common.h" has been kept so that old code can
still build with it, and so that it can be re-introduced in the future
if needed.