ap: Refactor AP connection request code
In the shared WifiData struct the current AP information has been split
from the password and wepmode variables. THis means we can just reuse the
same AP struct for everything.

Wifi_ConnectAP() has also been simplified. Instead of looking for the
requested AP in the list of found APs we ask Wifi_AssocStatus() to do it.
This means that there are only two places in the code that can set a new
AP to connect to.