ap: Simplify AP connection state machine
Previously, the ARM9 used flag WFLAG_REQ_APCONNECT of variable reqReqFlags
to ask the ARM7 to connect to an AP. This was orthogonal to reqMode, which
caused the code to be quite awkward in some places.

Now, whenever the ARM9 wants the ARM7 to connect to an AP, it sets reqMode
to WIFIMODE_ASSOCIATED, and it sets it to WIFIMODE_NORMAL when it wants
to disconnect. This behaviour is more similar to the rest of the modes.
For example, scan mode is entered by requesting WIFIMODE_SCAN. Access point
mode (for multiplayer hosts) is requested by requesting
WIFIMODE_ACCESSPOINT.