ap: Implement Wifi_ConnectSecureAP()
Wifi_ConnectAP() uses WEPMODE enums to determine the size of the key. This
isn't very convenient, and it makes it difficult to add support for other
encryption types.

Wifi_ConnectSecureAP() takes a key and a key length as input, which means
it can work for any type of key, and it does internally any kind of
required conversion for WEP.