lwip: Properly bring lwIP net interface up and down
Detect changes in the connection to the Access Point from the ARM7 and
communicate them to the ARM9 so that it can bring the lwIP netif up and
down.

This also makes it possible for lwIP to terminate active operations
right away when the network interface is brought down, preventing waits
until a timeout stops the operation.