ntr: Decouple curMode and authlevel state machines
Previously, some autthentication and association code modified the value
of WifiData->curMode directly, which made it hard to keep track of where
the code was changing its value.

WIFI_AUTHLEVEL_ERROR has been added to the enum of authentication levels
so that code outside of Wifi_NTR_Update() can use that value instead of
setting WifiData->curMode.

Also, this makes the code on the NTR driver behave a lot more like the
TWL driver, which doesn't have any coupling between WMI and curMode.