wpa: Support WPA2 APs when the user only provides the password
The ARM7 will detect if there is a valid PMK or not and it will calculate
it if not. This is a process that takes about 5 seconds.

When users connect to APs configured in the WFC settings this isn't needed
because the firmware stores the PMK in its settings, so the connection
happens right away.