twl: Refactor RX/TX queues
The new format consists of a u32 with the size of the data, the data padded
up to a 32-bit size, followed by another size. A CPU must always write a
a 32-bit 0 value after the data packet so that the other CPU stops there.
A 0 value means "no more packets to read", while 0xFFFFFFFF means "go back
to the beginning of the buffer".

The new sizes stored in TWL mode in WifiData->rxbufWrite/Read and
WifiData->txbufWrite/Read are in bytes so that it's easier to understand
the code without a bunch of "* 2" or "/ 2" everywhere.

Also, stop initializing the TX/RX buffers when we connect to an AP. This
isn't really needed because each CPU will naturally handle all packages
even if they aren't connected to the AP (the ARM7 just drops the packets,
the ARM9 will handle them and eventually the connection will fail when no
more packets arrive).