ntr: rx: Refactor RX IPC queue
Make it work like in TWL mode. The circular buffer (written by the ARM7)
now doesn't split packets when the end of the buffer is reached. Instead,
it jumps to the start so that the packet is always in one piece. This
allows some optimizations now, and it will allow more in the future.