rx: tx: Fix race condition when sending packets between CPUs
The first 32 bits written to the circular buffer indicate the size of the
packet that goes after it, but it also indicates that the packet is ready
to be read by the other CPU. We need to wait until the end to write to it.

It's very unlikely this was causing bugs because the other CPU normally
doesn't check the buffer until the FIFO sync event is handled, but it
could have happened eventually.