ntr: tx: Refactor code that sends packets to the ARM7
The type of transfer is specified in the 32-bits that are reserved for
the size in the circular buffer. The 4 top bits are reused as flags for
CMD/REPLY/DATA/BEACON packets.

Also, WIFI_TRANSFER_RATE_1MBPS and WIFI_TRANSFER_RATE_2MBPS have been made
public again because Wifi_RawTxFrame() requires them.