docs: ntr: Use DMA to read/write packets in MAC RAM
Using DMA is faster than the CPU, particularly to read packets because we
don't need to worry about checking for overflows while reading from MAC
RAM: The W_RXBUF_RD_ADDR register wraps around automatically.

Also, we don't need to worry about cache because we're copying from ARM7
RAM to main RAM, but in a buffer that the ARM9 only reads from an uncached
main RAM mirror.

The old code has been left in the codebase in case this causes issues
and it has to be reverted. However, it works on real hardware and in
melonDS.