rand: Implement new RNG system
This new system can be seeded at any point. The ARM7 is in charge of
seeding it from different sources (currently only the time and W_RANDOM in
DS mode). More sources need to be added later.

This seed isn't actually used to generate random numbers directly. It's
used to create seeds for the ARM7 and ARM9 to generate numbers.

The current system can definitely be improved, but it's already an
improvement over the previous system (rand() from stdlib.h is a pretty bad
RNG, and W_RANDOM isn't great either).