lwip: Use the right startup netif flag
We control the "link up" status to tell lwIP if the WiFi connection is
ready or not. The netif status shouldn't start as "link up", it should
wait until the code detects that the connection is ready.

However, we don't modify the "up" netif status, so we can set it to "up"
from the start.