ipc: Fix race conditions when writing packets that wrap around
If we write the wrap marker before setting the next size to 0 (at the base
of the TX/RX buffer) it's possible that the other CPU will see the wrap
marker, then read an invalid size, and keep going. This will eventually
lead to a crash due to reading a size that is too big.

Also, when we see a wrap marker we can't assume that the value at the base
of the RX/TX buffer is valid. We need to check that it isn't 0.